backend/pyproject.toml
[project]
name = "consensia-backend"
version = "0.1.0"
description = "FastAPI backend for orchestrating AI personas and judge consensus."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "pydantic>=2.9,<3.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "openai>=1.51.0,<2.0.0",
  "google-generativeai>=0.8.2,<1.0.0",
  "python-multipart>=0.0.7,<1.0.0",
  "pdfplumber>=0.11.0,<1.0.0",
  "python-docx>=1.0.0,<2.0.0",

  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "pyjwt>=2.8",
  "pwdlib[argon2]>=0.2",
  "google-auth>=2.35",
  "email-validator>=2.2",
  "google-search-results>=2.4.2"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.5,<1.0.0",
  "pytest>=8.3.0,<9.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]